% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS.UGS.R
\name{AGS.UGS}
\alias{AGS.UGS}
\title{AGS/UGS Merchandizing}
\usage{
AGS.UGS(Stand, Plot, Tree, SPP, DBH, HT, Stump, AGS.UGS = "UGS")
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species idientification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{Stump}{Stump height in meters. Recommended value if not measured is .5}

\item{AGS.UGS}{Tree is recorded as AGS or UGS, defaults to UGS}
}
\value{
data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF.AGS, Saw.Vol.AGS,
Pulp.Vol.AGS, Cull.Vol.AGS, Total.Vol, Merch.Vol, Percent.Sawlog.AGS)
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using AGS/UGS measurements.
}
\details{
AGS is acceptable growing stock. UGS is unacceptable growing stock.

AGS/UGS can have multiple definitions, but this function determines all
portions of the tree to be sawlog to minimum inside bark diameter, pulp
to minimumum inside bark diameter, and cull for the remainder for all 'AGS'
trees. All UGS trees will be calculated as CULL.

ALL TREES DEFAULT TO UGS WHEN NO VALUE PROVIDED

Volumes determined using Kozak Taper Equations and Smalians Volume Formula.
Merch diameters establish by the merc function.

df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into dataframe when used with mapply.
}
\examples{
AGS.UGS(1, 1, 1, 'RS', 30, 14, .5, 'AGS')
AGS.UGS(1, 1, 2, 'RO', 25, 12, .5, 'UGS')
AGS.UGS(1, 1, 3, 'SM', 40, 18, .5)

}
