% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASG.R
\name{ASG.USG}
\alias{ASG.USG}
\title{ASG Merchandising}
\usage{
ASG.USG(Stand, Plot, Tree, SPP, DBH, HT, Stump, GS = "USG")
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species identification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{HT}{Height of tree in meters}

\item{Stump}{Stump height in meters. Recommended value if not measured is .5}

\item{GS}{Record the growing stock of the tree as ASG or USG, defaults to USG}
}
\value{
Metric, with the exception of Board Feet which is returned with imperial values.
\subsection{}{

data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF.ASG, Saw.Vol.ASG,
Pulp.Vol.ASG, Cull.Vol.ASG, Total.Vol, Merch.Vol, Percent.Sawlog.ASG)
}
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using Acceptable or Unacceptable Sawlog Grade Observations Inputs should be in metric.
}
\details{
ASG is any tree tree with at least one 8 foot sawlog grade section.
USG trees have no sawlogs and are marked as either Pulp or Cull.

ASG/USG can have multiple definitions, but this function assumes all
portions of the tree to be sawlog to minimum outside bark merchantable diameter, pulp
to minimum outside bark merchantable diameter, and cull for the remainder for all 'ASG'
trees. All USG trees will be calculated as Pulp or CULL.
\subsection{All trees default to USG when no value is input.}{

Volumes determined using Kozak Taper Equations and Smalians Volume Formula.
Merch diameters establish by the MerchDiam function.

Sawlog board feet is estimated using the international 1/4 inch rule. The sawlog portion of the stem is broken
into 4 sections of equal length and the international 1/4 inch rule is applied to each section.

df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into dataframe when used with mapply.
}
}
\examples{
ASG.USG(1, 1, 1, 'RS', 30, 14, .5, 'ASG')
ASG.USG(1, 1, 2, 'RO', 25, 12, .5, 'USG')
ASG.USG(1, 1, 3, 'SM', 40, 18, .5)

}
\seealso{
\link{KozakTreeVol}

\link{KozakTaper}

\link{MerchDiam}

Other Merchandising Functions: 
\code{\link{Form.Risk}()},
\code{\link{MerchHT}()},
\code{\link{Sawlog.Likelihood}()},
\code{\link{Stick.Cruise}()},
\code{\link{ValueEstimate}()}
}
\author{
Ryan Smith
}
\concept{Merchandising Functions}
