% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalArea.R
\name{BA}
\alias{BA}
\title{Basal Area Formula}
\usage{
BA(DBH)
}
\arguments{
\item{DBH}{Diameter at breast height in inches}
}
\value{
BA - Basal Area in sq ft
}
\description{
This function calculates the basal area of a tree.
The basal area is the cross sectional area of the tree.
}
