% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAPH.R
\name{BAPH}
\alias{BAPH}
\title{Basal Area Per Hectre}
\usage{
BAPH(Stand, Plot, BA, EXPF)
}
\arguments{
\item{Stand}{The unique Stand ID for each Plot.}

\item{Plot}{The unique Plot ID for each Plot.}

\item{BA}{Diameter at breast height in cm.}

\item{EXPF}{Enter either the percent of an acre for fixed or BAF for variable.}
}
\value{
Returns a vector of the length N with the basal area in sq meters per hectare for each plot.
}
\description{
This function calculates the Basal Area Per Hectre for each plot within your dataset.
}
\details{
This function requires that you calculate the Basal Area and the Expansion Factor for each tree in
your inventory. See the EXP.F and BA functions if you need to do this.
\subsection{}{

Basal Area Per Hectre is a dependency for a number of other functions that may be useful in your analysis, so
using a standard nameing convention, while always good practice, will help when applying to additional functions.
(ex: BAPH, BA.PH, BasalAreaPerHectare, BAPerHectare).
}
}
\examples{
# Fixed Area - EXP.F(16, .2)
# Variable Radius - EXP.F(16, 20)


Stand <- (1,1,1,1)
Plot  <- (1,1,1,1)
BA    <- (.01, .12, .06, .04)
EXPF  <- (16, 16, 16, 16)
BAPH(Stand, Plot, BA, EXPF)
#
# tibble <- tibble \%>\% mutate(BAPH = BAPH(Stand, Plot, BA, EXPF))

}
\seealso{
\link{BA}

\link{EXP.F}

Other Basal Area Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BA}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{Unique.ID}()}
}
\concept{Basal Area Functions}
\concept{Plot Level Functions}
