% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasTesting.R
\name{Bias}
\alias{Bias}
\title{Bias Testing of Measured vs. Expected Results}
\usage{
Bias(estim, measu)
}
\arguments{
\item{estim}{Estimated Values}

\item{measu}{Measured Values}
}
\value{
"mean error", "mean absolute error", "RMSE", "Mean \% SE", "PB"
}
\description{
This function calculates the mean error, mean absolute error,
RMSE, mean percent standard error, and mean percent error using
observed values as denominator and not taking abs values.
}
\details{
The length of the estimated and measured values must be equal.
}
\author{
Aaron Weiskittel
}
