% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCFL.R
\name{CCF.Larger}
\alias{CCF.Larger}
\title{Crown Competition in Larger Trees}
\usage{
CCF.Larger(ID, SPP, DBH, EXPF)
}
\arguments{
\item{SPP}{Species for each tree}

\item{DBH}{Diameter at breast height in cm.}

\item{EXPF}{Expansion Factor for each tree. Can be calculated with the EXP.F function.}

\item{Unique.ID}{Unique Plot ID defined by the Unique.ID function}
}
\value{
The function returns CCF values of larger trees for each tree in a plot as a numeric vector of length n.
}
\description{
\subsection{This function requires that your data.frame be sorted as detailed in examples before running.}{

This function calculates the cumulative Crown Competition Factor of larger trees within a plot.
This function must be run with the same sorting function as the BAL function.
Please refer to details for explanation of outputs and required inputs, as this function
will not run without properly sorting your inventory as demonstrated in the details and/or examples.
The expansion factor for measured trees must be computed first using the EXP.F function.
}
}
\details{
This function uses the dplyr package to select, arrange, and mutate the data.
\subsection{This function requires that your data.frame be sorted as such prior to running:}{

Tibble <- Tibble \%>\%
group_by(ID) \%>\%
arrange(desc(DBH), .by_group = TRUE)

Your data should include unique IDs specific to each plot. If you have not done this, it can be done using the
Unique.ID function included in this package.
}

\subsection{}{

This function is a dependency for a variety of additional functions in the inventoryfunctions package,
and the example inventory script included in this package will demonstrate how to execute it prior to
running those additional dependencies. Thus, it is recommended that you run the function into an object
named BAL in your data.frame or tibble.
}

\subsection{}{
}
}
\examples{

# trees_2010 <- trees_2010 \%>\%
# group_by(ID) \%>\%
#  arrange(desc(DBH), .by_group = TRUE)

# trees_2010 <- trees_2010 \%>\%
# mutate(
#   CCFL = CCF.Larger(ID, SPP, CBH, EXPF)
# )

}
\seealso{
\link{Unique.ID}

\link{EXP.F}

\link{BA.Larger.Trees}

Other Values in Larger Trees: 
\code{\link{BA.Larger.Trees}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CrownCompF}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{Unique.ID}()}

Other Crown Functions: 
\code{\link{CrownCompF}()},
\code{\link{LCW}()},
\code{\link{MCW}()}
}
\concept{Crown Functions}
\concept{Plot Level Functions}
\concept{Values in Larger Trees}
