% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrownCompetitionFactor.R
\name{CrownCompF}
\alias{CrownCompF}
\title{Crown Competition Factor}
\usage{
CrownCompF(Stand, Plot, Tree, SPP, DBH, EXPF, CUTOFF = TRUE)
}
\arguments{
\item{Plot}{Unique Plot ID}

\item{Tree}{Unique Tree ID}

\item{SPP}{Tree Species: use the FVS code}

\item{DBH}{Diameter at breast height in cm.}

\item{EXPF}{Expansion factor for each tree.}

\item{CUTOFF}{The minimum diameter in cm of a tree to be included in the analysis. If left blank the default CUTOFF is 10cm.}
}
\value{
This function will return a numeric vector of the Crown Competition Factor for each plot in your inventory.
}
\description{
This function returns the plot level Crown Competition Factor (CCF). The CCF is based on
the MCW of trees per Hectare. By default, this function only includes trees with a DBH > 10
in the analysis. This DBH can be changed to any value by adding a CUTOFF value.
}
\details{
\subsection{Metric}{

This function uses metric units.

Species should be entered as a character vector. This function uses the dplyr package to select,
arrange, and mutate the data. It is important that
you enter vectors of equal length otherwise the package will give you an error message.
}

\subsection{}{

MCW value needed to calculate CCF, please refer to MCW function in see also.
}

\subsection{}{

CCF estimates the area available to the average tree in a stand relative
to the maximum area it could use if it were open-growth. Although not a measure
of crown closure (Curtis, 1970), complete crown closure of unthinned stands is
generally assumed to occur from a CCF of 100 to the maximum for the species (Krajicek et al., 1961),
which can be over 500 for some species like Douglas-fir and western hemlock (e.g. Hann et al., 2003).
}
}
\examples{

Stand <- c(1,1,1,1,1,1)
Plot  <- c(1,1,1,2,2,2)
Tree  <- c(1,2,3,1,2,3)
SPP   <- c("BF", "RO", "RS", "YB", "RO", "YB")
DBH   <- c(24, 34, 18, 41, 6, 20)
EXPF <- c(5, 5, 5, 5, 5, 5)
CrownCompF(Stand, Plot, Tree, SPP, DBH, EXPF)
CrownCompF(Stand, Plot, Tree, SPP, DBH, EXPF, CUTOFF = 2)
CrownCompF(Stand, Plot, Tree, SPP, DBH, EXPF, CUTOFF = 21)

# Tibble \%>\% mutate(
#    CCF = CrownCompF("Stand ID", Plot ID", "Tree ID", "SPP Variable",
#              "DBH Variable", "EXPF Variable")
#  )

}
\references{
Hann, David & Marshall, David & Hanus, Mark. (2003). Equations for predicting height-to-crown-base, 5-year diameter
growth rate, 5-year height growth rate, 5-year mortality rate and maximum size-density trajectory for Douglas-fir
and western hemlock in the coastal region of the Pacific Northwest. Oregon State University, Forest Research Laboratory,
Research Contribution 40.

Vezina, P. E. (1962). CROWN WIDTH-D.B.H. RELATIONSHIPS FOR OPEN-GROWN BALSAM FIR AND WHITE SPRUCE IN QUEBEC.
The Forestry Chronicle, 38(4), 463–473. https://doi.org/10.5558/tfc38463-4

Weiskittel, A.R., D.W. Hann, J.A. Kershaw Jr and J.K. Vanclay. 2011a. Forest growth and yield modeling.
Wiley. Chichester, UK.
}
\seealso{
\link{MCW}

\link{EXP.F}

Other Crown Functions: 
\code{\link{CCF.Larger}()},
\code{\link{LCW}()},
\code{\link{MCW}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{HeightPredict}()},
\code{\link{QMD}()},
\code{\link{RD}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{TallestTrees}()},
\code{\link{Unique.ID}()}
}
\concept{Crown Functions}
\concept{Plot Level Functions}
