% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOBtoDIB.R
\name{DOBtoDIB}
\alias{DOBtoDIB}
\title{Diameter OB to Diameter IB}
\usage{
DOBtoDIB(SPP, dob)
}
\arguments{
\item{SPP}{= Species: use FVS species codes}

\item{dob}{= diameter outside bark}
}
\description{
This function converts diameter measurement to inside bark volume when desired.
Uses outside bark diameter breast height diameter measurements.
This function is primarily used to compliment the Kozak Volume Function.
}
\details{
\subsection{}{

Metric
}

\subsection{}{

Bark = measurement inside or outside bark: ob = outside, ib = inside
}
}
\examples{
DOBtoDIB(SPP = 'RO', dob = 40)

}
\seealso{
\link{KozakTaper}

\link{KozakTreeVol}
}
