% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Form.Risk.R
\name{Form.Risk}
\alias{Form.Risk}
\title{Form.Risk Merchandizing}
\usage{
Form.Risk(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  HT,
  Form = "AF",
  Risk = "LR",
  Cull = FALSE
)
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species idientification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{HT}{Height of tree in meters}

\item{Form}{Form classes 1-8 or 'GF', 'AF', 'PF' values are accepted. Defaults to 'AF'}

\item{Risk}{Risk class may be entered using 1-4 values or 'HR' or 'LR'. Defaults to 'LR'}

\item{Cull}{if tree is cull enter TRUE, defaults to FALSE}
}
\value{
Metric, with the exception of Board Feet which is returned with imperial values.
\subsection{}{

data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF.FR, Saw.Vol.FR,
Pulp.Vol.FR, Cull.Vol.FR, Total.Vol, Merch.Vol, Percent.Sawlog.FR)
}
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using Form/Risk analysis.
}
\details{
For explanation of form/risk analysis see castle. et al 2017 and
the NHRI Silvicultural guide for New Brunswick.

Volumes determined using Kozak Taper Equations and Smalians Volume Formula.
Merch diameters establish by the MerchDiam function.

Sawlog board feet is estimated using the international 1/4 inch rule. The length of the stem that is sawlog
quality is calculated based on the predicted sawlog volume using the Kozak Taper Equation.
The The sawlog portion of the stem is then broken into 2.4384m sections of equal length and the
international 1/4 inch rule is applied to each section. If the final section is longer than
2.4384m but smaller than 4.8768m then that entire length will be used as the final log for
calculating board feet. \cr
object <- as.data.frame(object)\cr
df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)\cr
is a useful pipe for unnesting the lists into dataframe when used with mapply.

The only species this will currently execute for are Sugar Maple ("SM"),
Yellow Birch ("YB"), Red Maple ("RM"), and Red Oak ("RO").
}
\examples{
Form.Risk(1, 1, 1, 'YB', 30, 14, 1, 2)
Form.Risk(1, 1, 2, 'RO', 25, 12, 4, 4, TRUE)
Form.Risk(1, 1, 2, 'RO', 25, 12, 7, 2)
Form.Risk(1, 1, 3, 'SM', 40, 18, 'GF', 'LR')
Form.Risk(1, 1, 3, 'SM', 40, 18, 'PF', 'LR')
Form.Risk(1, 1, 3, 'SM', 40, 18, 'AF', 'HR', TRUE)
}
\seealso{
\link{KozakTreeVol}

\link{KozakTaper}

\link{MerchDiam}

Other Merchandising Functions: 
\code{\link{ASG}()},
\code{\link{MerchHT}()},
\code{\link{Sawlog.Likelihood}()},
\code{\link{Stick.Cruise}()},
\code{\link{ValueEstimate}()}
}
\author{
Ryan Smith
}
\concept{Merchandising Functions}
