% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LargestCrownWidth.R
\name{LCW}
\alias{LCW}
\title{Largest Crown Width}
\usage{
LCW(SPP, MCW, DBH)
}
\arguments{
\item{SPP}{Tree Species: use the FVS code}

\item{MCW}{Maximum Crown Width in meters.}

\item{DBH}{Diameter at breast height in cm.}
}
\value{
Numeric. Widest crown diameter or radius for a stand grown tree in meters.
}
\description{
This function returns an estimate for the widest crown diameter or radius for a stand grown tree.
}
\details{
\subsection{Metric}{

This function uses metric units.

SPP included: BF, BS, EH, WP, NC, RS, WS, AB, GB, RB, RO, PB, QA, RM, SM, YB.
SPP that are not included can be enter as 'OH' for hardwoods and 'OS' for softwoods.
If no SPP is entered a standard coefficient will be used but may not reflect the
crown conditions accurately - please enter 'OS' or 'OH' if species is not included.

Species should be entered as a character vector.

MCW value needed to calculate LCW, please refer to MCW function in see also.
}
}
\examples{

LCW("BF", 8.63, 25)
# Tibble \%>\% mutate(LCW = LCW("SPP Variable", "MCW Variable", "DBH Variable"))

}
\seealso{
\link{MCW}

Other Crown Functions: 
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{MCW}()}
}
\concept{Crown Functions}
