% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxCrownWidth.R
\name{MCW}
\alias{MCW}
\title{Maximum Crown Width}
\usage{
MCW(SPP, DBH)
}
\arguments{
\item{SPP}{Tree Species: use the FVS code}

\item{DBH}{Diameter at breast height in inches}
}
\value{
The maximum crown width for the tree.
}
\description{
This function returns the maximum crown width for an open grown tree
for the most common species found in the acadian forest.
}
\details{
SPP included: BF, BS, EH, WP, NC, RS, WS, AB, GB, RB, RO, PB, QA, RM, SM, YB.
SPP that are not included can be enter as 'OH' for hardwoods and 'OS' for softwoods.
If no SPP is entered a standard coefficent will be used but may not reflect the
crown conditions accurately - please enter 'OS' or 'OH' if species is not included.
}
