% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merc.R
\name{merc}
\alias{merc}
\title{Minimum Merchantable Diameter For Product by Species}
\usage{
merc(SPP)
}
\arguments{
\item{SPP}{Species: use FVS species codes}
}
\value{
saw, pallet, pulp
}
\description{
This function establishes the minimium merchantable diameter for sawlog,
pulp, and pallet products for each species. As these dimentions are constantly
changing and location dependent, these values only provide a baseline. If
values need to be changed for a specific use create a new function with the
changed values. If the species is not capable of producing a value the minimum
diameter is set at 1000. Diameters are all stated in cm.
}
\examples{
merc("RO")

}
