% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merchantable.Height.R
\name{Merchantable.Height}
\alias{Merchantable.Height}
\title{Merchantable Stopper Height Merchandizing}
\usage{
Merchantable.Height(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  HT,
  Stump,
  Saw.Height,
  Pulp = "OTHER"
)
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species idientification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{Stump}{Stump height in meters, can be value where sawlog quality stem begins}

\item{Saw.Height}{Height at which sawlog quality stem ends. Value is 0 for pulp of cull trees.}

\item{Pulp}{defaults to 'Other'. For pulp and cull trees value is
'Pulp' or 'Cull' and Saw.Height value is 0.}
}
\value{
data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF.MH, Saw.Vol.MH,
Pulp.Vol.MH, Cull.Vol.MH, Total.Vol, Merch.Vol, Percent.Sawlog.MH)
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using merchantable stopper height measurements.
In this method a beginning 'Stump' height and a top 'Saw.Height'
are measured along the portion of the stem that is determined to be
sawlog quality. All volumes calculated using Kozak.
Each section that is not entered defults to Cull.
Merch diameters establish by the merc function.
}
\details{
df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into dataframe when used with mapply.
}
\examples{
Merchantable.Height(1, 1, 1, 'RS', 30, 14, .5, 6)
Merchantable.Height(1, 1, 2, 'RS', 30, 14, .5, 0, 'Pulp')
Merchantable.Height(1, 1, 3, 'RS', 30, 14, .5, 0, 'Cull')

}
