% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MerchDiam.R
\name{MerchDiam}
\alias{MerchDiam}
\title{Minimum Merchantable Diameter For Product by Species}
\usage{
MerchDiam(SPP)
}
\arguments{
\item{SPP}{Species: use FVS species codes}
}
\value{
This function returns a vector with the minimum dimensions for sawlogs (saw),
pallet wood (pallet), and pulp (pulp) for the desired species.
}
\description{
This function establishes the minimum merchantable diameter for sawlog,
pulp, and pallet products for each species. As these dimensions are constantly
changing and location dependent, these values only provide a baseline.
If the species is not capable of producing a value the minimum
diameter is set at 1000. Diameters are all stated in cm.
}
\details{
\subsection{If merch diameters need to be changed for a specific use this function should be changed locally}{
}

\subsection{but should not be included in any update to this package.}{
}
}
\examples{
MerchDiam("RO")

}
