% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MerchHT.R
\name{MerchHT}
\alias{MerchHT}
\title{Merchantable Stopper Height Merchandising}
\usage{
MerchHT(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  HT,
  Stump,
  Saw.Height,
  Pulp = TRUE,
  Cull = FALSE
)
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species identification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{HT}{Tree height in meters}

\item{Stump}{Stump height in meters, can be value where sawlog quality stem begins}

\item{Saw.Height}{Height at which sawlog quality stem ends. Value is 0 for pulp of cull trees.}

\item{Pulp}{defaults to TRUE. If the tree is pulp quality, enter TRUE.}

\item{Cull}{defaults to FALSE. If the tree is cull quality, enter TRUE.}
}
\value{
This function will return the cubic volumes of product potential in cubic meters. All inputs are metric
and all outputs are metric with the exception of Board Feet which is returned with imperial values.
\subsection{}{

Returns the following dataframe:
}

\subsection{}{

data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF.MH, Saw.Vol.MH,
Pulp.Vol.MH, Cull.Vol.MH, Total.Vol, Merch.Vol, Percent.Sawlog.MH)
}
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using merchantable stopper height measurements.
In this method a beginning 'Stump' height and a top 'Saw.Height'
are measured along the portion of the stem that is determined to be
sawlog quality. All volumes calculated using Kozak Volume Function.
Merch diameters establish by the MerchDiam function.
}
\details{
This function will automatically downgrade products to pulp for portions of the
measured sawlog stem where minimum sawlog dimensions are not met. Only the sawlog portion of the
stem meeting minimum diameter requirements is used to estimate board feet.

Sawlog board feet is estimated using the international 1/4 inch rule. The sawlog portion of the
stem is broken into 2.4384m sections and the international 1/4 inch rule is applied
to each section. If the final section is longer than 2.4384m but smaller than 4.8768m
then that entire length will be used as the final log for calculating board feet.

df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into dataframe when used with mapply.
}
\examples{
MerchHT(1, 1, 1, 'RS', 30, 14, .5, 6)
MerchHT(1, 1, 1, 'SM', 52, 14, 2, 7)
MerchHT(1, 1, 2, 'RS', 41, 14, .5, 0, 'Pulp')
MerchHT(1, 1, 3, 'RS', 30, 14, .5, 0, 'Cull')

}
\seealso{
\link{KozakTreeVol}

\link{KozakTaper}

\link{MerchDiam}

Other Merchandising Functions: 
\code{\link{ASG.USG}()},
\code{\link{Form.Risk}()},
\code{\link{Sawlog.Likelihood}()},
\code{\link{Stick.Cruise}()},
\code{\link{ValueEstimate}()}
}
\author{
Ryan Smith
}
\concept{Merchandising Functions}
