% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMD.R
\name{QMD}
\alias{QMD}
\title{QMD}
\usage{
QMD(DBH, EXPF)
}
\arguments{
\item{DBH}{Diameter at breast height in cm.}

\item{EXP.F}{Expansion factor for each tree.}
}
\value{
Returns a vector of the length N with the quadratic mean diameter of each plot for each tree in your data set.
}
\description{
This function calculates the quadratic mean diameter value of each plot for every tree within your data set.
}
\details{
This function requires that you calculate the Basal Area Per Hectare (BAPH) and Trees Per Hectare (TPH) for every tree
in your dataset. Values must be metric and in hectares. These values can be calculated using the BAPH and TPH functions included in this
package.
}
\examples{

TPH <- c(1200, 1300, 2000, 700)
BAPH <- c(14, 20, 60, 12)
QMD(TPH, BAPH)

}
\seealso{
\link{EXP.F}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{HeightPredict}()},
\code{\link{RD}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{TallestTrees}()},
\code{\link{Unique.ID}()}
}
\author{
Ryan Smith
}
\concept{Plot Level Functions}
