% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeDensity.R
\name{RD}
\alias{RD}
\title{Relative Density}
\usage{
RD(SDIPlot, SDIMax)
}
\arguments{
\item{SDIPlot}{The calculated SDI at each plot.}

\item{SDIMax}{The calculated SDImax for each plot.}
}
\value{
The return value will be a relative density which is a ratio of Stand Density and Maximum Stand Density.
}
\description{
This function computes the Relative Density Index for each plot in your inventory.
}
\details{
Stand density is a quantitative measure of the degree of crowding
and resulting level of competition existing within the stand (Zeide, 2005).

Combined with the SDI.Plot (SDI using summation method) and SDI.Max (using average plot Standard Gravity)
functions, this is a measure for examining and comparing competition within mixed and uneven aged stands.
\subsection{Interpreting Relative Density}{

Based on their approach of estimating relative density, Drew and Flewelling (1979) suggest that relative densities
of 0.15, 0.40, and 0.55 correspond to the onset of competition, lower limit of full site occupancy, and the zone of
imminent competition mortality, respectively. Curtis (2010) suggests that trees less than 4 cm DBH should be excluded
from the computation of any relative density measure (Weiskittel et al. 2011, 25).
}

\subsection{Inputs - Plot, Stand, and Tree Tables}{

This function can be used in three different ways. First, you can run a tibble or dataframe of tree level
measurements where each tree has a corresponding SDI and SDImax based on its plot or stand location. Second,
you can run a tibble or datafram of plot level data that will provide you with the relative density ratios at each plot.
Third, you can utilize this at the stand level, by running mean stand density and mean stand density max data for an entire
stand.

The ratio is unit less so it can be used for both imperial and metric data sets.
}
}
\examples{

SDIPlot  <- c(1200, 987, 1823)
SDIMax   <- c(2100, 2050, 2150)
RD(SDIPlot, SDIMax)

}
\seealso{
\link{SDI.Tree}

\link{SDI.Plot}

\link{SDI.Max}

Other Stand Density Index Functions: 
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{SDI.Tree}()}
}
\concept{Stand Density Index Functions}
