% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sawlog.Likelihood.R
\name{Sawlog.Likelihood}
\alias{Sawlog.Likelihood}
\title{Likelihood of Having A Sawlog}
\usage{
Sawlog.Likelihood(SPP, DBH, Form, Risk)
}
\arguments{
\item{SPP}{The species identification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{Form}{Form classes 1-8 or 'GF', 'AF', 'PF' values are accepted.}

\item{Risk}{Risk class may be entered using 1-4 values or 'HR' or 'LR'.}
}
\value{
Returns a predictive value to be used in AUC analysis to determine the likelihood that a tree contains a sawlog.
}
\description{
For explanation of form/risk analysis see Castle. et al 2017 . This function
is meant to supplement the Form.Risk merchandising function.
}
\details{
The only species this will currently execute for are Sugar Maple ("SM"),
Yellow Birch ("YB"), Red Maple ("RM"), and Red Oak ("RO").
\subsection{}{

These values can be used in a multitude of ways, but if you are looking for binary TRUE/FALSE results
for if a tree is likely to produce a sawlog, the optimal cutoff point for determining whether or not a tree
has a sawlog can be found using the coords function
in the pRoc package. Cutpoints that have been used with exceptionally high AUC values (greater than .85)
in Northern Maine are:\cr
Red Maple: Cutpoint 0.737935\cr
Sugar Maple: Cutpoint 0.8500554\cr
Yellow Birch: Cutpoint 0.8586569\cr
}
}
\examples{
Sawlog.Likelihood("RO", 42, 1, 1)
}
\seealso{
\link{Form.Risk}

Other Merchandising Functions: 
\code{\link{ASG.USG}()},
\code{\link{Form.Risk}()},
\code{\link{MerchHT}()},
\code{\link{Stick.Cruise}()},
\code{\link{ValueEstimate}()}
}
\author{
Ryan Smith
}
\concept{Merchandising Functions}
