% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.R
\name{SDI}
\alias{SDI}
\title{Single Tree SDI Function}
\usage{
SDI(DBH)
}
\arguments{
\item{DBH}{Diameter at Breast Height}
}
\description{
This function computes the SDI value of a single tree with
no regard for its expansion factor or the influence of the tree
on a plot or stand level. See Details.
}
\details{
This function is meant to be applied on a plot or stand level using
the included EXP.F function and dplyr. See the included full inventory script
for its application on a stand and plot level.

For example: an effective use of this function would be:

Tree \%>\% mutate(SDI = SDI(DBH)*EXP.F)
}
