% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalians.R
\name{smalians}
\alias{smalians}
\title{Smalians volume function}
\usage{
smalians(r1, r2, len)
}
\arguments{
\item{r1}{diameter on one end of cylinder}

\item{r2}{diameter on second end of cylinder}

\item{len}{length of cylinder}
}
\description{
This function compliments the kozak volume function.
Function provides volume of a cylindrical object.
}
