% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stick.Cruise.R
\name{Stick.Cruise}
\alias{Stick.Cruise}
\title{Stick Cruise Merchandising}
\usage{
Stick.Cruise(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  HT,
  S1 = "Pulp",
  S2 = "Pulp",
  S3 = "Pulp",
  S4 = "Pulp",
  S5 = "Pulp",
  S6 = "Pulp",
  S7 = "Pulp",
  S8 = "Pulp",
  S9 = "Pulp",
  S10 = "Pulp",
  Cull = FALSE,
  Sawlogs = FALSE,
  StartingLog = FALSE,
  UnlimLog = FALSE,
  Veneer = FALSE
)
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species identification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{HT}{Tree height in meters}

\item{S1}{Each section called out as 'Saw', 'Pulp', or 'Cull'.}

\item{Cull}{if it is a cull tree than enter TRUE, else enter FALSE. Default is FALSE.}

\item{Sawlogs}{If each individual section is not called, you may use this
input to call the total number of Sawlogs in the tree. If individual sections are called
no Sawlog argument should be entered.}

\item{StartingLog}{If the 8 foot section where sawlogs begin is recorded you can use
StartingLog to name which section to begin your Sawlogs input at. This can only be used
with the Sawlogs parameter and should not be used if individual sections are called.}

\item{UnlimLog}{If all sections that meet diameter requirements
are sawlogs enter UnlimLog as TRUE.}

\item{S2...}{You can call as many as 10 'S#' sections.}
}
\value{
Metric, with the exception of Board Feet which is returned with imperial values.
\subsection{}{

data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF, Saw.Vol,
Pulp.Vol, Cull.Vol, Total.Vol, Merch.Vol, Percent.Sawlog)
}
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using stick cruise measurements. In a stick cruise each 2.4384 meter section of the tree
is called out as either potential saw, pulp, or cull. Up to 19 sections may be entered.
See details for input options.
}
\details{
Sawlogs may be entered in two different fashions.
\subsection{}{

The first option is to simply enter the total number of sawlogs called.
When the argument Sawlogs = # is entered the function will call that
number of 2.4384 meter sections of the stem as Sawlog grade beginning at a default
stump height of .5 meters. The maximum number of Sawlogs that can be called is 10. You
can set a section to begin your sawlog count at using the StartingLog param. This is useful
if the first 2.4384 meter section of the stem has disqualifying defect. If no value is
entered for a StartingLog then the measurements will begin at the .5m stump.
}

\subsection{}{

The second option is to call each 2.4384 meter individual section of the tree. The maximum number
of section calls is 10. Beginning with S1, each section may be entered as "Saw", "Pulp",
or "Cull". You may call as few sections as you would like, and each uncalled merchantable section of the stem
will default to "Pulp". The default stump height is .5 meters.
}

\subsection{}{

Using the second input option will give you a more accurate volume and sawlog board foot
estimate if there is a "Pulp" or "Cull" call before or between Sawlog calls.
If this option is used no argument should be entered for the Sawlog or StartingLog params.
}

\subsection{}{

This function will automatically downgrade products to pulp for sections where a
sawlog product call is made but where the small end inside bark diameter of the section does
not meet minimum sawlog dimensions for the species. Merch dimensions can be altered and species can be added
by altering the internal MerchDiam function. Outside bark diameters are measured using the Kozak Taper Equation.
}

\subsection{}{

Sawlog board feet is estimated using the international 1/4 inch rule. The international 1/4 inch rule is applied
to each eight foot section that is called as "Saw" and which meets diameter requirements
}

\subsection{}{

df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into a dataframe when this function
is used for a large number of trees using the mapply function.
}
}
\examples{
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, S1 = 'Saw', S2 = 'Pulp', S3 = 'Saw')
Stick.Cruise.Tree(1, 2, 2, 'RO', 68, 26, 'Pulp', 'Saw', 'Saw', Pulp,
                 S5 = "Saw", S6 = "Saw", S7 = "Pulp" S8 = "Cull")
Stick.Cruise.Tree(857, 9, 16, 'SM', 62, 24, S3 = "Saw", S5 = "Saw")
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, Sawlogs = 2)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, Sawlogs = 2, StartingLog = 2)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, Cull = TRUE)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16)
Stick.Cruise.Tree(1, 1, 1, 'AB', 30, 16, 'Saw', 'Pulp', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RS', 30, 16, 'Pulp', 'Saw', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RS', 30, 16, Sawlogs = 2)
Stick.Cruise.Tree(1, 1, 1, 'AB', 30, 16, Sawlogs = 5, StartingLog = 3)

}
\seealso{
\link{KozakTreeVol}

\link{KozakTaper}

\link{MerchDiam}

Other Merchandising Functions: 
\code{\link{ASG}()},
\code{\link{Form.Risk}()},
\code{\link{MerchHT}()},
\code{\link{Sawlog.Likelihood}()},
\code{\link{ValueEstimate}()}
}
\author{
Ryan Smith
}
\concept{Merchandising Functions}
