% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stick.Cruise.Tree.R
\name{Stick.Cruise.Tree}
\alias{Stick.Cruise.Tree}
\title{Stick Cruise Merchandizing}
\usage{
Stick.Cruise.Tree(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  HT,
  Cull = FALSE,
  S1 = "Pulp",
  S2 = "Pulp",
  S3 = "Pulp",
  S4 = "Pulp",
  S5 = "Pulp",
  S6 = "Pulp",
  S7 = "Pulp",
  S8 = "Pulp",
  S9 = "Pulp",
  S10 = "Pulp",
  S11 = "Pulp",
  S12 = "Pulp",
  S13 = "Pulp",
  S14 = "Pulp",
  S15 = "Pulp",
  S16 = "Pulp",
  S17 = "Pulp",
  S18 = "Pulp",
  S19 = "Pulp",
  S20 = "Pulp"
)
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species identification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{HT}{Tree height in meters}

\item{Cull}{if it is a cull tree than enter TRUE, else enter FALSE}

\item{S1...}{Each section called out as 'Saw', 'Pulp', or 'Cull'.}
}
\value{
data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF.SC, Saw.Vol.SC,
Pulp.Vol.SC, Cull.Vol.SC, Total.Vol, Merch.Vol, Percent.Sawlog.SC)
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using stick cruise measurements. In a stick cruise each 2.4384 meter section of the tree
is called out as either potential saw, pulp, or cull. Up to 20 sections may be entered.
First section begins at .5 meters.
Each section that is not entered defaults to Pulp.
Merch diameters establish by the merc function.
}
\details{
df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into dataframe when used with mapply.
}
\examples{
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, FALSE, 'Saw', 'Pulp', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, TRUE, 'Saw', 'Pulp', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, FALSE)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, TRUE)

}
