% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TallestTrees.R
\name{TallestTrees}
\alias{TallestTrees}
\title{Average Height of the 100 Tallest Trees}
\usage{
TallestTrees(ID, HT, EXPF)
}
\arguments{
\item{ID}{Unique Plot ID for each plot in your data set.}

\item{HT}{Height of each tree in meters.}

\item{EXPF}{Expansion factor of each tree.}
}
\value{
The function returns a vector of length n with the 100 tallest trees in each plot.
}
\description{
\subsection{This function requires that your data.frame be sorted as detailed in examples before running.}{

This function calculates the average height of the 100 tallest trees in your data set. Using the EXPF and heights provided
the 100 tallest trees will be identified and averaged. This mean value will then be applied to a variable listing the average
of the tallest trees within the plot as an observation for each tree in your data set.
}
}
\details{
This function uses the dplyr package to select, arrange, and mutate the data.
\subsection{This function requires that your data frame be sorted as such prior to running:}{

df <- df \%>\%
group_by(ID) \%>\%
arrange(desc(HT), .by_group = TRUE)

Your data should include unique IDs specific to each plot. If you have not done this, it can be done using the
Unique.ID function included in this package. This function requires either a measured or predicted height for each
tree in your inventory.
}
}
\examples{

\dontrun{
data(tree_data)
tree_data$ID <- Unique.ID(tree_data$Stand, tree_data$Plot)
tree_dat$EXPF <- EXP.F(DBH, PlotSize)

tree_data <- tree_data \%>\%
group_by(ID) \%>\%
arrange(desc(HT), .by_group = TRUE)

tree_data <- tree_data \%>\%
mutate(
   TallestTrees = TallestTrees(ID, HT, EXPF)
 )

tree_data$TallestTrees
}


}
\seealso{
\link{Unique.ID}

\link{EXPF}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{HeightPredict}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{Unique.ID}()}
}
\concept{Plot Level Functions}
