% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_data_documentation.R
\docType{data}
\name{tree_data}
\alias{tree_data}
\title{A Sample Tree Inventory}
\format{
A data frame with 277 rows and 11 variables:
\describe{
\item{Stand}{Unique Stand ID}
\item{Plot}{Unique Plot ID}
\item{Tree}{Unique Tree ID}
\item{SPP}{Tree Species Code}
\item{DBH}{Diameter at Breast Height in CM}
\item{Saw Height}{The observed SawHeight in the sub sample where HT measurements were taken}
\item{HT}{The observed HT measurement in the sub sample where HT measurements were taken}
\item{Form}{The identified tree Form Class based on NHRI Form Classes}
\item{Risk}{The identified tree Ris Class based on NHRI Form Classes}
\item{X}{Longitude}
\item{Y}{Latitude}
...
}
}
\usage{
tree_data
}
\description{
This datasets contains 277 tree observations along with standard tree level measurements and observations.
}
\keyword{datasets}
