% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unique.Id.R
\name{Unique.ID}
\alias{Unique.ID}
\title{Unique Plot ID Code}
\usage{
Unique.ID(Stand, Plot)
}
\arguments{
\item{Stand}{The unique stand ID for each stand.}

\item{Plot}{The unique plot ID for each plot within a stand.}
}
\value{
This function returns a vector of length n with unique id
numbers for every individual plot.
}
\description{
When working with a large data set of multiple stands and plots,
it can be helpful to provide each individual plot a unique ID code.
As an example, his ensures that when using some functions,
like basal area of larger trees, you don't confused plot 2 in
stands 1, 2, 3.
}
\examples{
Stand <- c(1,1,1,1,1,1,2,2,2,2,2,2)
Plot  <- c(1,2,3,4,5,6,1,2,3,4,5,6)
Unique.ID(Stand, Plot)

}
\seealso{
Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()}
}
\concept{Plot Level Functions}
