% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueEstimate.R
\name{ValueEstimate}
\alias{ValueEstimate}
\title{Value Estimate}
\usage{
ValueEstimate(Stand, Plot, Tree, SPP, BoardFeet, PulpVol)
}
\arguments{
\item{Stand}{Stand ID for Plot where the nth tree is located.}

\item{Plot}{Plot ID for Plot where the nth tree is located.}

\item{Tree}{The ID for the tree.}

\item{SPP}{Tree Species using FVS SPP codes.}

\item{BoardFeet}{Etimated Sawlog Boardfeet for the tree.}

\item{PulpVol}{The estimated volume as pulp in cubic meters}
}
\value{
Returns a list of Sawlog and Pulp values for each sample tree.
}
\description{
This function estimates the Sawlog and Pulp values of each sample tree based
on the estimated sawlog BF and pulp volume. Value estimates are based on
Average Maine 2019 Statewide stumpage prices.
}
\details{
Each tree is given a value based on the average price paid per MBF or ib for pulp.
If no value is reported either in the statewide report
the product will be deemed non-commercial and given a value of $0.

It is recommended that Stand and Plot be factors and each plot must have a unique ID
and not be a recycled number (see details section of the HeightPredict function).
}
\examples{

ValueEstimate(1, 1, 1, "SM", 382.48, 1.19)

}
\seealso{
\link{HeightPredict}

Other Merchandising Functions: 
\code{\link{ASG}()},
\code{\link{Form.Risk}()},
\code{\link{MerchHT}()},
\code{\link{Sawlog.Likelihood}()},
\code{\link{Stick.Cruise}()}
}
\author{
Ryan Smith
}
\concept{Merchandising Functions}
