% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAL.R
\name{BA.Larger.Trees}
\alias{BA.Larger.Trees}
\title{Basal Area in Larger Trees}
\usage{
BA.Larger.Trees(ID, DBH, BA)
}
\arguments{
\item{DBH}{Diameter at breast height in cm.}

\item{BA}{Basal area of each tree in sq meters. Can be calculated with the BA function.}

\item{Unique.ID}{Unique Plot ID defined by the Unique.ID function}
}
\value{
The function returns BAL values in sq. meters as a numeric vector of length n.
}
\description{
\subsection{This function requires that your data.frame be sorted as detailed in examples before running.}{

This function calculates the cumulative basal areas of larger trees within a plot.
This function must be run with the same sorting function as the CCFL function.
Please refer to details for explanation of outputs and required inputs, as this function
will not run without properly sorting your inventory as demonstrated in the details and/or examples.
The basal area for measured trees must be computed first using the BA function.
}
}
\details{
This function uses the dplyr package to select, arrange, and mutate the data.
\subsection{This function requires that your data frame be sorted as such prior to running:}{

df <- df \%>\%
group_by(ID) \%>\%
arrange(desc(DBH), .by_group = TRUE)

Your data should include unique IDs specific to each plot. If you have not done this, it can be done using the
Unique.ID function included in this package.
}

\subsection{}{

This function is a dependency for a variety of additional functions in the inventoryfunctions package,
and the example inventory script included in this package will demonstrate how to execute it prior to
running those additional dependencies. Thus, it is recommended that you run the function into an object
named BAL in your data.frame or tibble.
}

\subsection{}{
}
}
\examples{

# trees_2010 <- trees_2010 \%>\%
# group_by(ID) \%>\%
#   arrange(desc(DBH), .by_group = TRUE)

# trees_2010 <- trees_2010 \%>\%
# mutate(
#    BAL = BA.Larger.Trees(ID, DBH, BA)
#  )

}
\seealso{
\link{BA}

\link{Unique.ID}

\link{CCF.Larger}

Other Values in Larger Trees: 
\code{\link{CCF.Larger}()}

Other Basal Area Functions: 
\code{\link{BAPH}()},
\code{\link{BA}()}

Other Plot Level Functions: 
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{TallestTrees}()},
\code{\link{Unique.ID}()}
}
\concept{Basal Area Functions}
\concept{Plot Level Functions}
\concept{Values in Larger Trees}
