% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalArea.R
\name{BA}
\alias{BA}
\title{Basal Area Formula}
\usage{
BA(DBH)
}
\arguments{
\item{DBH}{Diameter at breast height in cm.}
}
\value{
BA - Basal Area in sq meters
}
\description{
This function calculates the basal area of a tree.
The basal area is the cross sectional area of the tree.
\subsection{}{

Inputs and return values are metric.
}
}
\examples{

BA(34)

# Tibble \%>\% mutate(BA = BA("DBH Vector"))

}
\seealso{
Other Basal Area Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()}
}
\concept{Basal Area Functions}
