% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalArea.R
\name{BA}
\alias{BA}
\title{Basal Area Formula}
\usage{
BA(DBH)
}
\arguments{
\item{DBH}{Diameter at breast height in cm.}
}
\value{
A numeric vector with the basal area of n trees.
}
\description{
This function calculates the basal area of a tree in square meters.
The basal area is the cross sectional area of the tree.
\subsection{}{

Inputs and return values are metric.
}
}
\examples{

BA(34)

\dontrun{
# Tibble \%>\% mutate(BA = BA("DBH Vector"))
}
}
\references{
Kershaw, J. A., Ducey, M. J., Beers, T. W., & Husch, B. (2017).
Forest mensuration (Fifth edition). Wiley/Blackwell.
}
\seealso{
Other Basal Area Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()}
}
\concept{Basal Area Functions}
