% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAL.R
\name{BAL}
\alias{BAL}
\title{Basal Area in Larger Trees}
\usage{
BAL(Plot, Tree, BA)
}
\arguments{
\item{Plot}{Unique Plot ID}

\item{Tree}{Unique Tree ID within each plot}

\item{BA}{Basal area of each tree in sq meters. Can be calculated with the BA function.}
}
\value{
Numeric. The function returns BAL values in sq. meters as a vector of length n.
}
\description{
This function calculates the cumulative basal areas of larger trees within a plot.
Please refer to details for explanation of outputs and required inputs.
The basal area for measured trees must be computed first using the BA function.
}
\details{
This function uses the dplyr package to select, arrange, and mutate the data. It is important that
you enter vectors of equal length otherwise the package will give you an error message.
\subsection{}{

Your data should include individual plot IDs for each plot otherwise the BAL value will be inaccurate.
}

\subsection{}{

Metric values as inputs and returns.
}

\subsection{}{

This function is a dependency for a variety of additional functions in the inventoryfunctions package,
and the example inventory script included in this package will demonstrate how to execute it prior to
running those additional dependencies. Thus, it is recommended that you run the function into an object
named BAL in your dataframe or tibble. A recommended way to run the function is this:
}

\subsection{}{

Tibble <- Tibble \%>\% mutate(BAL = BAL(Plot, Tree, BA))
}

\subsection{}{

If you prefer base R, just make sure it is saved as the variable BAL in your df so that it can be easily integrated
into additional functions for which it is a dependency.
}
}
\examples{

Plot <- c(1,1,1,2,2,2)
Tree <- c(1,2,3,1,2,3)
BA   <- c(.04, .12, .06, .11, .12, .06)
BAL(Plot, Tree, BA)

}
\seealso{
\link{BA}
}
