% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAPH.R
\name{BAPH}
\alias{BAPH}
\title{Basal Area Per Hectare}
\usage{
BAPH(Stand, Plot, BA, EXPF)
}
\arguments{
\item{Stand}{The unique Stand ID for each Plot.}

\item{Plot}{The unique Plot ID for each Plot.}

\item{BA}{Diameter at breast height in cm.}

\item{EXPF}{Expansion factor for each tree.}
}
\value{
Returns a vector of the length N with the basal area in sq meters per hectare for each plot.
}
\description{
This function calculates the Basal Area Per Hectare for each plot within your data set.
}
\details{
This function requires that you calculate the Basal Area and the Expansion Factor for each tree in
your inventory. See the EXP.F and BA functions if you need to do this.
\subsection{}{

Basal Area Per Hectare is a dependency for a number of other functions that may be useful in your analysis, so
using a standard naming convention, while always good practice, will help when applying to additional functions.
(ex: BAPH, BA.PH, BasalAreaPerHectare, BAPerHectare).
}
}
\examples{

Stand <- c(1,1,1,1)
Plot  <- c(1,1,1,1)
BA    <- c(.01, .12, .06, .04)
EXPF  <- c(16, 16, 16, 16)
BAPH(Stand, Plot, BA, EXPF)

}
\seealso{
\link{BA}

\link{EXP.F}

Other Basal Area Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BA}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{HeightPredict}()},
\code{\link{QMD}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{TallestTrees}()},
\code{\link{Unique.ID}()}
}
\concept{Basal Area Functions}
\concept{Plot Level Functions}
