% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCFL.R
\name{CCF.Larger}
\alias{CCF.Larger}
\title{Crown Competition in Larger Trees}
\usage{
CCF.Larger(ID, SPP, DBH, EXPF, CUTOFF = TRUE)
}
\arguments{
\item{SPP}{Species for each tree}

\item{DBH}{Diameter at breast height in cm.}

\item{EXPF}{Expansion Factor for each tree. Can be calculated with the EXP.F function.}

\item{CUTOFF}{The minimum diameter in cm of a tree to be included in the analysis. If left blank the default CUTOFF is 10cm.}

\item{Unique.ID}{Unique Plot ID defined by the Unique.ID function}
}
\value{
The function returns CCF values of larger trees for each tree in a plot as a numeric vector of length n.
}
\description{
\subsection{This function requires that your data.frame be sorted as detailed in examples before running.}{

This function calculates the cumulative Crown Competition Factor of larger trees within a plot.
This function must be run with the same sorting function as the BA.Larger function.
Please refer to details for explanation of required inputs, as this function
will not run without properly sorting your inventory as demonstrated in the details and/or examples.
}
}
\details{
This function calculates the crown competition factor in larger trees. By default, this function will only
include trees with a DBH > 10 in the analysis. This value can be increased or decreased by adding a cutoff value.
\subsection{This function requires that your data.frame be sorted as such prior to running:}{

df <- df \%>\%
group_by(ID) \%>\%
arrange(desc(DBH), .by_group = TRUE)

Your data should include unique IDs specific to each plot. If you have not done this, it can be done using the
Unique.ID function included in this package. Each tree must have an expansion factor value. This can be done using
the EXP.F function.
}
}
\examples{

ID <- c(1,1,1,1,1)
SPP <- c("BF", "RO", "RM", "BF", "RO")
DBH <- c(25, 30, 8, 19, 32)
EXPF <- c(8, 8, 8, 8, 8)
CCF.Larger(ID, SPP, DBH, EXPF, CUTOFF = TRUE)
CCF.Larger(ID, SPP, DBH, EXPF, CUTOFF = 5)

\dontrun{
df <- df \%>\%
   dplyr::group_by(ID) \%>\%
   dplyr::arrange(desc(DBH), .by_group = TRUE)

df <- df \%>\%
   dplyr::mutate(
   CCFL = CCF.Larger(ID, SPP, DBH, EXPF)
 )
}

}
\seealso{
\link{Unique.ID}

\link{EXP.F}

\link{BA.Larger.Trees}

Other Values in Larger Trees: 
\code{\link{BA.Larger.Trees}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CrownCompF}()},
\code{\link{HeightPredict}()},
\code{\link{QMD}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{TallestTrees}()},
\code{\link{Unique.ID}()}

Other Crown Functions: 
\code{\link{CrownCompF}()},
\code{\link{LCW}()},
\code{\link{MCW}()}
}
\concept{Crown Functions}
\concept{Plot Level Functions}
\concept{Values in Larger Trees}
