% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrownCompetitionFactor.R
\name{CrownCompF}
\alias{CrownCompF}
\title{Crown Competition Factor}
\usage{
CrownCompF(Stand, Plot, Tree, SPP, DBH, EXPF)
}
\arguments{
\item{Plot}{Unique Plot ID}

\item{Tree}{Unique Tree ID}

\item{SPP}{Tree Species: use the FVS code}

\item{DBH}{Diameter at breast height in cm.}

\item{EXPF}{Expansion factor for each tree.}
}
\value{
This function will return the Crown Competition Factor for each plot in your inventory.

This return will give you a CCF based on the MCW of trees per Hectare.
}
\description{
This function returns the plot level Crown Competition Factor (CCF).
}
\details{
\subsection{Metric}{

This function uses metric units.
}

\subsection{}{

Species should be entered as a character vector. This function uses the dplyr package to select,
arrange, and mutate the data. It is important that
you enter vectors of equal length otherwise the package will give you an error message.
}

\subsection{}{

MCW value needed to calculate CCF, please refer to MCW function in see also.
}

\subsection{}{

CCF estimates the area available to the average tree in a stand relative
to the maximum area it could use if it were open-growth. Although not a measure
of crown closure (Curtis, 1970), complete crown closure of unthinned stands is
generally assumed to occur from a CCF of 100 to the maximum for the species (Krajicek et al., 1961),
which can be over 500 for some species like Douglas-fir and western hemlock (e.g. Hann et al., 2003).
}
}
\examples{

Stand <- c(1,1,1,1,1,1)
Plot  <- c(1,1,1,2,2,2)
Tree  <- c(1,2,3,1,2,3)
SPP   <- c("BF", "RO", "RS", "YB", "RO", "YB")
DBH   <- c(24, 34, 18, 41, 28, 20)
EXPF <- c(5, 5, 5, 5, 5, 5)
CrownCompF(Stand, Plot, Tree, SPP, DBH, EXPF)


# Tibble \%>\% mutate(
#    CCF = CrownCompF("Stand ID", Plot ID", "Tree ID", "SPP Variable",
#              "DBH Variable", "EXPF Variable")
#  )

}
\seealso{
\link{MCW}

\link{EXP.F}

Other Crown Functions: 
\code{\link{CCF.Larger}()},
\code{\link{LCW}()},
\code{\link{MCW}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Plot}()},
\code{\link{Unique.ID}()}
}
\concept{Crown Functions}
\concept{Plot Level Functions}
