% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOBtoDIB.R
\name{DOBtoDIB}
\alias{DOBtoDIB}
\title{Diameter Outside Bark to Diameter Inside Bark}
\usage{
DOBtoDIB(SPP, dob)
}
\arguments{
\item{SPP}{Species: use FVS species codes: example 'RO' - Red Oak, 'WS' = White Spruce}

\item{dob}{Outside bark diameter of a location on the tree.}
}
\description{
This function converts the outside bark diameter measurement (cm) of a tree to the diameter inside bark (cm).
Useful for inside bark volumes or for making conversions to set minimal merchantable diamters in
FVS which only allows for IB diameters.
}
\details{
All diameters must be metric (cm). Includes species found in the acadian forest.
}
\examples{
DOBtoDIB("SM", 25.4)
DOBtoDIB("RS", 15)
DOBtoDIB("PB", 12)

}
\seealso{
\link{EXP.F}

Other Volume Functions: 
\code{\link{KozakTreeVol}()}
}
\concept{Volume Functions}
