% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Form.Risk.R
\name{Form.Risk}
\alias{Form.Risk}
\title{Form.Risk Merchandizing}
\usage{
Form.Risk(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  HT,
  Stump,
  Form = "AF",
  Risk = "LR",
  Cull = FALSE
)
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species idientification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{Stump}{Stump height in meters. Recommended value if not measured is .5}

\item{Form}{Form classes 1-8 or 'GF', 'AF', 'PF' values are accepted. Defaults to 'AF'}

\item{Risk}{Risk class may be entered using 1-4 values or 'HR' or 'LR'. Defaults to 'LR'}

\item{CULL}{if tree is cull enter TRUE, defaults to FALSE}
}
\value{
data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF.FR, Saw.Vol.FR,
Pulp.Vol.FR, Cull.Vol.FR, Total.Vol, Merch.Vol, Percent.Sawlog.FR)
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using Form/Risk analysis.
}
\details{
For explaination of form/risk analsis see castle. et al 2017 and
the NHRI Silvicultural guide for New Brunswick.

Volumes determined using Kozak Taper Equations and Smalians Volume Formula.
Merch diameters establish by the merc function.

df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into dataframe when used with mapply.
}
\examples{
Form.Risk(1, 1, 1, 'RS', 30, 14, .5, 1, 2)
Form.Risk(1, 1, 2, 'RO', 25, 12, .5, 4, 4, TRUE)
Form.Risk(1, 1, 2, 'RO', 25, 12, .5, 7, 2)
Form.Risk(1, 1, 3, 'SM', 40, 18, .5, 'GF', 'LR')
Form.Risk(1, 1, 3, 'SM', 40, 18, .5, 'PF', 'LR')
Form.Risk(1, 1, 3, 'SM', 40, 18, .5, 'AF', 'HR', TRUE)
}
