% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genus.R
\name{Genus}
\alias{Genus}
\title{Returns Genus and Common Name for SPP in Inventory}
\usage{
Genus(SPP)
}
\arguments{
\item{SPP}{Species: use FVS species codes}
}
\value{
The 1: Common Name, 2: Genus 3: Lazy Name and
4: WoodType (HW/SW) of species entered.
}
\description{
This function returns the Genus, Wood Type (HW/SW), and Common Name for species
included in the Northeast Variant of FVS. Genus can be useful for
modeling purposes, and the common and lazy names can be used for creating
graphic representations of forest composition for non-foresters.
}
\examples{
trees <- c(1,2,3,4,5)
SPP <- c("BF", "BF", "RO", "RS", "RO")
tree_df <- data.frame(trees, SPP)
genusinfo <- apply(tree_df, 2, Genus)

Genus("BA")

}
