% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.Max.R
\name{SDI.Max}
\alias{SDI.Max}
\title{Plot Level Maximum Stand Density Function}
\usage{
SDI.Max(Stand, Plot, Tree, SPP, EXPF)
}
\arguments{
\item{Stand}{Unique Stand ID}

\item{Plot}{Unique Plot ID}

\item{Tree}{Unique Tree ID}

\item{SPP}{Diameter at breast height in cm.}

\item{EXPF}{Expansion factor for each tree.}
}
\value{
The return value will be a maximum stand density per hectare.
This function, along with \link{SDI.Plot}, is required to
compute the relative density of a stand with \link{RD}.
}
\description{
This function computes the Maximum Stand Density of each plot in your inventory based on the
standard gravity of the tree weighted by its expansion factor. This is a method of obtaining
Maximum Stand Density for uneven aged mixed stands.
}
\details{
Woodalls SDIMax
}
\examples{

Stand <- c(1,1,1,1,1,1)
Plot  <- c(1,1,1,2,2,2)
Tree  <- c(1,2,3,1,2,3)
SPP   <- c("RO", "WP", "EH", "YB", "YB", "SM")
EXPF <- c(746.037, 282.52, 86.45, 94.31, 165.21, 361.03)
SDI.Max(Stand, Plot, Tree, SPP, EXPF)

}
\seealso{
\link{SDI.Tree}

\link{SDI.Plot}

\link{RD}

Other Stand Density Index Functions: 
\code{\link{RD}()},
\code{\link{SDI.Plot}()},
\code{\link{SDI.Tree}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{Unique.ID}()}
}
\concept{Plot Level Functions}
\concept{Stand Density Index Functions}
