% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.Max.R
\name{SDI.Max}
\alias{SDI.Max}
\title{Plot Level Maximum Stand Density Function}
\usage{
SDI.Max(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  EXPF,
  CSI = NULL,
  X_Coord = NULL,
  Y_Coord = NULL,
  project = NULL,
  SpatialPointsDF = NULL
)
}
\arguments{
\item{Stand}{Unique Stand ID}

\item{Plot}{Unique Plot ID}

\item{Tree}{Unique Tree ID}

\item{SPP}{Species observation for every tree (FVS species code)}

\item{EXPF}{Expansion factor for each tree.}

\item{CSI}{Climate site index for each tree.}

\item{X_Coord}{Plot X coordinate for each tree}

\item{Y_Coord}{Plot Y coordinate for each tree}

\item{project}{The projection used for your provided coordinates will be EPSG:4326 unless another value is provided.}

\item{SpatialPointsDF}{A SpatialPointsDF may be provided for the data instead of X and Y coords.}
}
\value{
The return a numeric vector of length n with the value of the maximum stand density per hectare.
This function, along with SDI.Plot, is required to
compute the relative density of a stand with RD.
}
\description{
This function computes the Maximum Stand Density of each plot using one of two methods. If you include the necessary
Climate Site Index data and plot coordinates (lat, long or a SpatialPointsDF) SDImax will be calculated using the Weiskittel and Kuehne method.
If you fail to input this information SDImax will be calculated using Woodall's method based solely on
mean standard gravity.
}
\details{
Both methods in the description are proposed methods for obtaining Maximum Stand Density for uneven aged
mixed stands. Please see references for further information. It is the recommendation of the author of this
function that you utilize the Weiskittel method for calculating SDI.Max.

You can enter either a set of X and Y coordinates or a SpatialPointsDF into this function.
}
\examples{

Stand <- c(1,1,1,1,1,1)
Plot  <- c(1,1,1,2,2,2)
Tree  <- c(1,2,3,1,2,3)
SPP   <- c("RO", "WP", "EH", "YB", "YB", "SM")
DBH <- c(42, 23, 61, 25, 55, 15)
EXPF <- c(866.1473, 2888.249, 410.6111, 2444.614, 505.0856, 6790.595)
SDI.Max(Stand, Plot, Tree, SPP, DBH, EXPF)
}
\references{
Weiskittel, A. R., & Kuehne, C. (2019). Evaluating and modeling variation in site-level maximum carrying
capacity of mixed-species forest stands in the Acadian Region of northeastern North America.
The Forestry Chronicle, 95(03), 171–182. https://doi.org/10.5558/tfc2019-026

Woodall, C.W., P.D. Miles and J.S. Vissage. 2005. Determining maximum stand density index
in mixed species stands for strategic- scale stocking assessments. For. Ecol. Manage. 216: 367–377.
}
\seealso{
\link{SDI.Plot}

\link{RD}

Other Stand Density Index Functions: 
\code{\link{RD}()},
\code{\link{SDI.Plot}()}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{HeightPredict}()},
\code{\link{RD}()},
\code{\link{SDI.Plot}()},
\code{\link{TPH}()},
\code{\link{TallestTrees}()},
\code{\link{Unique.ID}()}
}
\author{
Ryan Smith
}
\concept{Plot Level Functions}
\concept{Stand Density Index Functions}
