% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.Plot.R
\name{SDI.Plot}
\alias{SDI.Plot}
\title{Plot Level Stand Density Index Function}
\usage{
SDI.Plot(Plot, Tree, DBH, EXP.F)
}
\arguments{
\item{Plot}{Unique Plot ID}

\item{Tree}{Unique Tree ID}

\item{DBH}{Diameter at breast height in cm.}

\item{EXP.F}{Expansion factor for each tree.}
}
\value{
The return value will be a stand density index for each plot in your
inventory. This SDI value is valuable when paired with the SDI.Max function to
obtain a Relative Density value using the RD function.
}
\description{
This function computes the Stand Density Index for each plot in your inventory.
}
\details{
This function utilizes the SDI.tree function and each trees expansion factor to
provide you with the SDI for each plot in your inventory.

SDI is obtained through summation.

This is a function simplifies obtaining plot level
SDI measurements as opposed to utilizing the example
code shown in SDI.Tree.
}
\examples{

Plot  <- c(1,1,1,2,2,2)
Tree  <- c(1,2,3,1,2,3)
DBH   <- c(24, 34, 18, 41, 28, 20)
EXP.F <- c(5, 5, 5, 5, 5, 5)
SDI.Plot(Plot, Tree, DBH, EXP.F)

}
\seealso{
\link{SDI.Tree}

\link{SDI.Max}

\link{RD}

Other Stand Density Index Functions: 
\code{\link{RD}()},
\code{\link{SDI.Max}()},
\code{\link{SDI.Tree}()}
}
\concept{Stand Density Index Functions}
