% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDI.Plot.R
\name{SDI.Plot}
\alias{SDI.Plot}
\title{Plot Level Stand Density Index Function}
\usage{
SDI.Plot(Stand, Plot, Tree, DBH, EXPF)
}
\arguments{
\item{Stand}{Unique Stand ID}

\item{Plot}{Unique Plot ID}

\item{Tree}{Unique Tree ID}

\item{DBH}{Diameter at breast height in cm.}

\item{EXPF}{Expansion factor for each tree.}
}
\value{
The return value is a numerical vector of length n providing the
stand density index for each plot in your inventory.
}
\description{
This function computes the Stand Density Index for each plot in your inventory using the summation method.
}
\details{
This function utilizes the internal SDI.tree function and each tree's expansion factor to
provide you with the SDI for each plot in your inventory.
\subsection{}{

The SDI.Tree function uses the formula (DBH/25.4)^1.605
}

\subsection{}{

This index of SDI is obtained through summation (see references). This is a way of determining stand density
in uneven-aged stands with non-normal diameter distributions.
}

\subsection{}{

This function simplifies obtaining plot level
SDI measurements. This SDI value is
useful when paired with the SDI.Max function to
obtain a Relative Density value using the RD function.
}
}
\examples{

Stand <- c(1,1,1,1,1,1)
Plot  <- c(1,1,1,2,2,2)
Tree  <- c(1,2,3,1,2,3)
DBH   <- c(24, 34, 18, 41, 28, 20)
EXPF <- c(5, 5, 5, 5, 5, 5)
SDI.Plot(Stand, Plot, Tree, DBH, EXPF)


}
\references{
Woodall, C. W., Miles, P. D., & Vissage, J. S. (2005). Determining maximum stand density index
in mixed species stands for strategic-scale stocking assessments.
Forest Ecology and Management, 216(1–3), 367–377. https://doi.org/10.1016/j.foreco.2005.05.050
}
\seealso{
\link{SDI.Tree}

\link{SDI.Max}

Other Plot Level Functions: 
\code{\link{BA.Larger.Trees}()},
\code{\link{BAPH}()},
\code{\link{CCF.Larger}()},
\code{\link{CrownCompF}()},
\code{\link{HeightPredict}()},
\code{\link{RD}()},
\code{\link{SDI.Max}()},
\code{\link{TPH}()},
\code{\link{TallestTrees}()},
\code{\link{Unique.ID}()}

Other Stand Density Index Functions: 
\code{\link{RD}()},
\code{\link{SDI.Max}()}
}
\author{
Ryan Smith
}
\concept{Plot Level Functions}
\concept{Stand Density Index Functions}
