% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smalians.R
\name{Smalians}
\alias{Smalians}
\title{Smalians volume function}
\usage{
Smalians(r1, r2, len)
}
\arguments{
\item{r1}{diameter on one end of cylinder}

\item{r2}{diameter on second end of cylinder}

\item{len}{length of cylinder}
}
\description{
This function compliments the Kozak Volume Function.
Function provides volume of a cylindrical object.
}
\seealso{
\link{KozakTreeVol}

Other Volume Functions: 
\code{\link{KozakTreeVol}()}
}
\concept{Volume Functions}
