% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stick.Cruise.Tree.R
\name{Stick.Cruise.Tree}
\alias{Stick.Cruise.Tree}
\title{Stick Cruise Merchandising}
\usage{
Stick.Cruise.Tree(
  Stand,
  Plot,
  Tree,
  SPP,
  DBH,
  HT,
  S1 = "Pulp",
  S2 = "Pulp",
  S3 = "Pulp",
  S4 = "Pulp",
  S5 = "Pulp",
  S6 = "Pulp",
  S7 = "Pulp",
  S8 = "Pulp",
  S9 = "Pulp",
  S10 = "Pulp",
  S11 = "Pulp",
  S12 = "Pulp",
  S13 = "Pulp",
  S14 = "Pulp",
  S15 = "Pulp",
  S16 = "Pulp",
  S17 = "Pulp",
  S18 = "Pulp",
  S19 = "Pulp",
  S20 = "Pulp",
  Stump = 0.5,
  Cull = FALSE,
  Sawlogs = FALSE,
  Veneer = FALSE
)
}
\arguments{
\item{Stand}{The Unique Stand Identification Number}

\item{Plot}{The Unique Plot Identification Number}

\item{Tree}{The Unique Tree Identification Number}

\item{SPP}{The species identification using FVS codes: ex 'RO' = Red Oak}

\item{DBH}{Diameter at breast height in cm}

\item{HT}{Tree height in meters}

\item{S1}{Each section called out as 'Saw', 'Pulp', or 'Cull'.}

\item{Cull}{if it is a cull tree than enter TRUE, else enter FALSE}

\item{Sawlogs}{If each individual section is not called, you may use this
input to call the total number of Sawlogs in the tree. If individual sections are called
no Sawlog argument should be entered.}

\item{S2...}{You can call as many as 20 sections.}
}
\value{
Metric, with the exception of Board Feet which is returned with imperial values.
\subsection{}{

data.frame(Stand, Plot, Tree, Method, SPP, Saw.BF, Saw.Vol,
Pulp.Vol, Cull.Vol, Total.Vol, Merch.Vol, Percent.Sawlog)
}
}
\description{
This function calculates the total tree volume, merchantable volume,
sawlog volume, pulp volume, cull volume, and saw board feet for trees
using stick cruise measurements. In a stick cruise each 2.4384 meter section of the tree
is called out as either potential saw, pulp, or cull. Up to 19 sections may be entered.
See details for input options.
}
\details{
Sawlogs may be entered in two different fashions.
\subsection{}{

The first option is to simply enter the total number of sawlogs called.
When the argument Sawlogs = # is entered the function will call that
number of 2.4384 meter sections of the stem as Sawlog grade beginning at a default
atump height of .5 meters. The maximum number of Sawlogs that can be called is 19.
This option assumes Sawlogs will be in the lower sections of the stem.
}

\subsection{}{

The second option is to call each 2.4384 meter individual section of the tree. The maximum number
of section calls is 19. Beginning with S1, each section may be entered as "Saw", "Pulp",
or "Cull". You may call as few sections as you would like, and each uncalled merchantable section of the stem
will default to "Pulp". The default stump height is .5 meters.
}

\subsection{}{

Using the second input option will give you a more accurate volume and sawlog board foot
estimate if there is a "Pulp" or "Cull" call before or between Sawlog calls.
If this option is used no argument should be entered for the Sawlog parameter.
}

\subsection{}{

This function will automatically downgrade products to pulp for sections where a
sawlog product call is made but where the small end inside bark diameter of the section does
not meet minimum sawlog dimensions for the species. Merch dimensions can be altered and species can be added
by altering the internal MerchDiam function. Inside bark diameters are measured using the Kozak Taper Equation.
}

\subsection{}{

Sawlog board feet is estimated using the international 1/4 inch rule. The international 1/4 inch rule is applied
to each eight foot section that is called as "Saw" and which meets diameter requirements
}

\subsection{}{

df <- df \%>\% rownames_to_column()
\%>\% gather(variable, value, -rowname) \%>\% spread(rowname, value)
is a useful pipe for unnesting the lists into a dataframe when this function
is used for a large number of trees using the mapply function.
}
}
\examples{
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, 'Saw', 'Pulp', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, 'Pulp', 'Saw', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, Sawlogs = 2)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, Cull = True)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, Sawlogs = 5)
Stick.Cruise.Tree(1, 1, 1, 'AB', 30, 16, 'Saw', 'Pulp', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RS', 30, 16, 'Pulp', 'Saw', 'Saw')
Stick.Cruise.Tree(1, 1, 1, 'RS', 30, 16, Sawlogs = 2)
Stick.Cruise.Tree(1, 1, 1, 'RO', 30, 16, Cull = True)
Stick.Cruise.Tree(1, 1, 1, 'QA', 30, 16)
Stick.Cruise.Tree(1, 1, 1, 'AB', 30, 16, Sawlogs = 5)

}
\seealso{
\link{KozakTreeVol}

\link{KozakTaper}

\link{MerchDiam}

Other Merchandising Functions: 
\code{\link{AGS.UGS}()},
\code{\link{Form.Risk}()},
\code{\link{Merchantable.Height}()},
\code{\link{Sawlog.Likelihood}()}
}
\author{
Ryan Smith
}
\concept{Merchandising Functions}
